+++ conans/client/importer.py
@@ -62,12 +62,14 @@ def _report_save_manifest(copied_files, output, dest_folder, manifest_name):
         manifest = FileTreeManifest(date, file_dict)
         save(os.path.join(dest_folder, manifest_name), str(manifest))
 
+
 def _make_files_writable(file_names):
     if not get_env("CONAN_READ_ONLY_CACHE", False):
         return
 
     for file_name in file_names:
         os.chmod(file_name, os.stat(file_name).st_mode | stat.S_IWRITE)
+
 
 def run_imports(conanfile, dest_folder, output):
     if not hasattr(conanfile, "imports"):
