+++ conans/test/functional/deps_graph_test.py
@@ -1,5 +1,5 @@
 import unittest
-from conans.client.deps_builder import DepsGraph, Node
+from conans.client.graph.graph_builder import DepsGraph, Node
 from conans.model.ref import ConanFileReference
 from conans.model.conan_file import ConanFile
 from conans.model.settings import Settings
